+++ src/ArduinoJson/Operators/VariantComparisons.hpp
@@ -4,9 +4,16 @@
 
 #pragma once
 
-#include <ArduinoJson/Variant/VariantRef.hpp>
+#include <ArduinoJson/Configuration.hpp>
+#include <ArduinoJson/Misc/Visitable.hpp>
+#include <ArduinoJson/Numbers/Float.hpp>
+#include <ArduinoJson/Numbers/Integer.hpp>
+#include <ArduinoJson/Polyfills/type_traits.hpp>
+#include <ArduinoJson/Strings/IsString.hpp>
 
 namespace ARDUINOJSON_NAMESPACE {
+
+class CollectionData;
 
 template <typename T, typename Enable = void>
 struct Comparer;