+++ src/ArduinoJson/Variant/VariantFunctions.hpp
@@ -4,6 +4,7 @@
 
 #pragma once
 
+#include <ArduinoJson/Polyfills/attributes.hpp>
 #include <ArduinoJson/Variant/VariantData.hpp>
 
 namespace ARDUINOJSON_NAMESPACE {